Update version number to v1.7.1.13.   Fix sign-extension bugs in ZipDirEntry and ZipEntry related to zip64.  Fix an overflow bug in the WinFormsApp relating to the progressBar value.
git-svn-id: https://dotnetzip.svn.codeplex.com/svn@27410 4ef1bfeb-0a34-4d02-9305-240e5b07a168

+++ Zip Partial DLL/ZipEntry.cs
@@ -860,8 +860,9 @@ namespace Ionic.Zip
 
                         i = 0;
                         ze._Crc32 = (Int32)(block[i++] + block[i++] * 256 + block[i++] * 256 * 256 + block[i++] * 256 * 256 * 256);
-                        ze._CompressedSize = block[i++] + block[i++] * 256 + block[i++] * 256 * 256 + block[i++] * 256 * 256 * 256;
-                        ze._UncompressedSize = block[i++] + block[i++] * 256 + block[i++] * 256 * 256 + block[i++] * 256 * 256 * 256;
+			ze._CompressedSize = (uint)(block[i++] + block[i++] * 256 + block[i++] * 256 * 256 + block[i++] * 256 * 256 * 256);
+			ze._UncompressedSize = (uint)(block[i++] + block[i++] * 256 + block[i++] * 256 * 256 + block[i++] * 256 * 256 * 256);
+
                     }
 
                     wantMore = (SizeOfDataRead != ze._CompressedSize);
